chore: remove helpers module
