+++ test/support/nx_signal_case.ex
@@ -1,5 +1,6 @@
 defmodule NxSignal.Case do
   use ExUnit.CaseTemplate
+  import ExUnit.Assertions
 
   using opts do
     validate_doc_metadata = Keyword.get(opts, :validate_doc_metadata, true)
@@ -34,6 +35,27 @@ defmodule NxSignal.Case do
         is_map(docs) and map_size(docs) > 0,
         metadata[:type] not in @doctypes do
       flunk("invalid @doc type: #{inspect(metadata[:type])} for #{name}/#{arity}")
+    end
+  end
+
+  @doc """
+  Asserts `lhs` is close to `rhs`.
+  """
+  def assert_all_close(lhs, rhs, opts \\ []) do
+    atol = opts[:atol] || 1.0e-4
+    rtol = opts[:rtol] || 1.0e-4
+
+    if Nx.all_close(lhs, rhs, atol: atol, rtol: rtol, equal_nan: opts[:equal_nan]) !=
+         Nx.tensor(1, type: {:u, 8}) do
+      flunk("""
+      expected
+
+      #{inspect(lhs)}
+
+      to be within tolerance of
+
+      #{inspect(rhs)}
+      """)
     end
   end
 end
